requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.7.0
Bootstrap-Flask==2.3.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
cycler==0.12.1
Django==5.0.3
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-environ==0.11.2
django-js-asset==2.2.0
django-recaptcha==4.0.0
django-utils-six==2.0
docopt==0.6.2
Flask==3.0.2
Flask-WTF==1.2.1
fonttools==4.46.0
future==0.18.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pipreqs==0.4.13
plotly==5.22.0
pyparsing==3.1.1
pyperclip==1.8.2
pypi==2.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scipy==1.11.4
seaborn==0.13.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tenacity==8.3.0
tzdata==2023.4
uncertainties==3.1.7
urllib3==2.1.0
Werkzeug==3.0.1
whitenoise==6.7.0
WTForms==3.1.2
yarg==0.1.9
